web/src/app/app.component.spec.ts
// src/app/app.component.spec.ts
import { TestBed } from '@angular/core/testing';
import {
  HttpClientTestingModule,
  HttpTestingController,
} from '@angular/common/http/testing';
import { AppComponent } from './app.component';
import { CurrencyService } from './service/currency.service'; // Assuming the file path is correct
import { of } from 'rxjs';

describe('AppComponent', () => {
  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [AppComponent],
      providers: [CurrencyService],
    }).compileComponents();
  });

  it('should create the app', () => {
    const fixture = TestBed.createComponent(AppComponent);
    const app = fixture.componentInstance;
    expect(app).toBeTruthy();
  });

  it(should have the 'proj' title, () => {
    const fixture = TestBed.createComponent(AppComponent);
    const app = fixture.componentInstance;
    expect(app.title).toEqual('proj');
  });

  it('should render title', () => {
    const fixture = TestBed.createComponent(AppComponent);
    fixture.detectChanges();
    const compiled = fixture.nativeElement as HTMLElement;
    expect(compiled.querySelector('h1')?.textContent).toContain('Hello, proj');
  });

  it('should fetch currency rate', () => {
    const fixture = TestBed.createComponent(AppComponent);
    const app = fixture.componentInstance;
    const currencyService = TestBed.inject(CurrencyService);
    const mockRate = { baseCurrency: 'USD', targetCurrency: 'EUR', rate: 0.85 };

    spyOn(currencyService, 'getRate').and.returnValue(of(mockRate));

    app.getCurrencyRate('USD', 'EUR');

    expect(app.rate).toEqual(mockRate);
  });
});